server: Pre-render Shapefiles to a Bytes representation
This is necessary primarily to reduce the per-request overhead.  We now
need only to read from the referenced segment of memory during response
time.  There is probably a better way to do this with a `static` type,
but this approach works quite well.

This is significantly more efficient, especially for large shapefiles.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>